Minimal documentation and move entityIdIndexer.js
