+++ app.js
@@ -6,7 +6,7 @@ const morgan = require('morgan');
 const config = require('config');
 const database = require('./model/database');
 const EntityIdIndexerModel = require('./model/EntityIdIndexer');
-const EntityIdIndexer = require('./middleware/entityIdIndexer')
+const EntityIdIndexer = require('./services/entityIdIndexer')
 
 const defaultPort = config.get('port');
 const port = process.env.PORT || defaultPort;
